+++ NEWS.md
@@ -1,5 +1,8 @@
 # dbplyr (development version)
 
+* Using a function with a namespace in `across()` now works, e.g.
+  `across(x, dplyr::dense_rank)` (@mgirlich, #1231).
+
 * The dots in `db_copy_to()` are now passed to `db_write_table()` (@mgirlich, #1237).
 
 * Can now use `select()` again after using `arrange(desc(x))` (@ejneer, #1240).